Fixed import statements after move to lib folder
